Adding check mode for delete operations for V3 and V4 modules (#652)
* check mode for delete operations

* response to msg fix

* response to msg fix

* add tests

* black fix

* fix for check mode

* Added test for check_mode delete operation ntnx_foundation_central

* syntax fixes

* Adding check mode changes in dev code

* adding check mode to tests & code refactoring

* black fix

* black fix

* Adding more check mode tests

* resolving comments

* black fix

* Ansible lint fixes

* Adding more black&list fixes

* flake8 fix

* Resolving comments

* Resolving comments

* black fix

* resolving comments

* Moving check mode to last in many tests

* resolving comment

* black fix

---------

+++ plugins/modules/ntnx_karbon_clusters_node_pools.py
@@ -319,6 +319,12 @@ def delete_nodes_of_pool(module, result):
 def delete_pool(module, result):
     cluster_name = module.params["cluster_name"]
     pool_name = module.params["node_pool_name"]
+    result["cluster_name"] = cluster_name
+    result["node_pool_name"] = pool_name
+
+    if module.check_mode:
+        result["msg"] = "Pool with name:{0} will be deleted.".format(pool_name)
+        return
 
     delete_nodes_of_pool(module, result)
 
@@ -326,8 +332,6 @@ def delete_pool(module, result):
     resp = node_pool.remove_node_pool(cluster_name, pool_name)
 
     result["changed"] = True
-    result["cluster_name"] = cluster_name
-    result["node_pool_name"] = pool_name
     task_uuid = resp["task_uuid"]
     result["task_uuid"] = task_uuid
