+++ plugins/modules/ntnx_ndb_profiles.py
@@ -1185,6 +1185,11 @@ def delete_profile(module, result):
     if not uuid:
         return module.fail_json(msg="'profile_uuid' is a required for deleting profile")
 
+    result["uuid"] = uuid
+    if module.check_mode:
+        result["msg"] = "Profile with uuid:{0} will be deleted.".format(uuid)
+        return
+
     resp = profiles.delete(uuid)
 
     result["response"] = resp
